(Ab)Use User::newSystemUser to disable accounts
While we're not actually turning the user into a "system" user, it has
the same end result: all passwords and other authentication credentials
removed or set to something invalid, email blanked, token invalidated,
and existing sessions dropped. So let's just use that if possible
instead of duplicating all the code, since "all the code" has become
somewhat more complex with the introduction of SessionManager and
AuthManager.

Bug: T110765
Change-Id: Ifce387d702860bca5ae9cdd90b660627258ff17e